requirements.txt
streamlit==1.27.2
python-dotenv==1.0.0
openai==1.12.0
typing_extensions>=4.5.0
pydantic>=2.0.0